+++ src/main/java/com/ifrs/edu/br/poll/repository/VoteRepository.java
@@ -3,5 +3,5 @@ package com.ifrs.edu.br.poll.repository;
 import com.ifrs.edu.br.poll.model.Poll;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VoteRepository extends JpaRepository<Poll, Long> {
+public interface VoteRepository extends JpaRepository<Poll, Integer> {
 }
